requirements.txt
Django==5.2.4
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
google-generativeai==0.8.5
crewai
python-dotenv
langchain-google-genai